refactor: Extract map-related functions to helpers file
