+++ app/charts/map/map.tsx
@@ -1,4 +1,3 @@
-import { WebMercatorViewport } from "@deck.gl/core";
 import { GeoJsonLayer, ScatterplotLayer } from "@deck.gl/layers";
 import { MapboxLayer } from "@deck.gl/mapbox";
 import { Box, Button, Theme } from "@mui/material";
@@ -6,8 +5,8 @@ import { makeStyles } from "@mui/styles";
 import { geoArea } from "d3";
 import { orderBy } from "lodash";
 import maplibregl from "maplibre-gl";
-import React, { useMemo, useRef, useState } from "react";
-import ReactMap, { LngLatLike, MapRef, ViewState } from "react-map-gl";
+import React, { useMemo, useRef } from "react";
+import ReactMap, { LngLatLike, MapRef } from "react-map-gl";
 
 import {
   emptyStyle,
@@ -18,154 +17,21 @@ import { useMapTooltip } from "@/charts/map/map-tooltip";
 import { convertHexToRgbArray } from "@/charts/shared/colors";
 import { useChartState } from "@/charts/shared/use-chart-state";
 import { useInteraction } from "@/charts/shared/use-interaction";
-import { Bounds } from "@/charts/shared/use-width";
 import { BBox } from "@/configurator/config-types";
 import { GeoFeature, GeoPoint } from "@/domain/data";
 import { Icon, IconName } from "@/icons";
-import useEvent from "@/lib/use-event";
 import { useLocale } from "@/src";
 
 import "maplibre-gl/dist/maplibre-gl.css";
 
+import { useViewState } from "./helpers";
 import Layer from "./layer";
 import { setMap } from "./ref";
-
-type MinMaxZoomViewState = Pick<
-  ViewState,
-  "zoom" | "latitude" | "longitude"
-> & {
-  minZoom: number;
-  maxZoom: number;
-  width: number;
-  height: number;
-};
-
-const BASE_VIEW_STATE: MinMaxZoomViewState = {
-  minZoom: 1,
-  maxZoom: 13,
-  latitude: 46.8182,
-  longitude: 8.2275,
-  zoom: 5,
-  width: 400,
-  height: 400,
-};
-
-/**
- * Compute map center along with a proper zoom level taking chart dimensions
- * into account.
- *
- * @param bbox Bounding box of the feature to be contained.
- * @param chartDimensions Chart's dimensions.
- */
-const getViewStateFromBounds = ({
-  bbox,
-  chartDimensions,
-}: {
-  bbox: BBox;
-  chartDimensions: Bounds;
-}) => {
-  const vp = new WebMercatorViewport({
-    ...BASE_VIEW_STATE,
-    width: chartDimensions.width,
-    height: chartDimensions.height,
-  });
-  const fitted = vp.fitBounds(bbox);
-
-  return { ...BASE_VIEW_STATE, ...fitted };
-};
-
-/**
- * Constrain the viewState to always _contain_ the supplied bbox.
- *
- * (Other implementations ensure that the bbox _covers_ the viewport)
- *
- * @param viewState deck.gl viewState
- * @param bbox Bounding box of the feature to be contained
- */
-const constrainZoom = (
-  viewState: MinMaxZoomViewState,
-  bbox: BBox,
-  { padding = 0 }: { padding?: number } = {}
-) => {
-  const vp = new WebMercatorViewport(viewState);
-
-  const { width, height, zoom, longitude, latitude } = viewState;
-
-  // Make sure the map is rendered before trying to project & fitBounds
-  if (vp.width > 1 && vp.height > 1) {
-    const [x, y] = vp.project([longitude, latitude]);
-    const [x0, y1] = vp.project(bbox[0]);
-    const [x1, y0] = vp.project(bbox[1]);
-
-    const fitted = vp.fitBounds(bbox, { padding });
-
-    const [cx, cy] = vp.project([fitted.longitude, fitted.latitude]);
-
-    const h = height - padding * 2;
-    const w = width - padding * 2;
-
-    const h2 = h / 2;
-    const w2 = w / 2;
-
-    const y2 =
-      y1 - y0 < h ? cy : y - h2 < y0 ? y0 + h2 : y + h2 > y1 ? y1 - h2 : y;
-    const x2 =
-      x1 - x0 < w ? cx : x - w2 < x0 ? x0 + w2 : x + w2 > x1 ? x1 - w2 : x;
-
-    const p = vp.unproject([x2, y2]);
-
-    return {
-      ...viewState,
-      zoom: Math.min(viewState.maxZoom, Math.max(zoom, fitted.zoom)),
-      longitude: p[0],
-      latitude: p[1],
-    };
-  } else {
-    return viewState;
-  }
-};
 
 let globalGeoJsonLayerId = 0;
 let globalScatterplotLayerId = 0;
 
 const FLY_TO_DURATION = 500;
-
-const useViewState = ({
-  chartDimensions,
-  bbox,
-  locked,
-}: {
-  chartDimensions: Bounds;
-  bbox: BBox | undefined;
-  locked: boolean;
-}) => {
-  const [viewState, setViewState] = useState(
-    locked && bbox
-      ? getViewStateFromBounds({ bbox, chartDimensions })
-      : BASE_VIEW_STATE
-  );
-
-  const onViewStateChange = useEvent(
-    ({ viewState }: { viewState: ViewState }) => {
-      setViewState((oldViewState) => ({ ...oldViewState, ...viewState }));
-    }
-  );
-
-  // Needed to be able to "refresh" the map to its initial position.
-  const initialViewState = useMemo(() => {
-    let newViewState: MinMaxZoomViewState | undefined = undefined;
-
-    if (bbox && !locked) {
-      newViewState = constrainZoom(viewState, bbox);
-      setViewState(newViewState);
-    }
-
-    return newViewState ?? viewState;
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  return { initialViewState, viewState, onViewStateChange };
-};
 
 export const MapComponent = () => {
   const locale = useLocale();
